finished project

+++ db/migrate/20240229230940_create_discounts.rb
@@ -1,7 +1,7 @@
 class CreateDiscounts < ActiveRecord::Migration[7.1]
   def change
     create_table :discounts do |t|
-      t.integer :percent_discount
+      t.float :percent_discount
       t.integer :quantity_threshold
       t.references :merchant, null: false, foreign_key: true
